break up runtime into multiple headers and source files
